Add Github Workflows and Setup for Documentation
