+++ tsconfig.json
@@ -65,5 +65,12 @@
   },
   "ts-node": {
     "files": true
+  },
+  "typedocOptions": {
+    "entryPoints": ["src/index.ts"],
+    "out": "docs",
+    "githubPages": true,
+    "includeVersion": true,
+    "customTitle": "CodeVachon / Next Version"
   }
 }
